[01-FizzBuzz, 02-TextMunger] Integration tests moved into default package.
