+++ 01-FizzBuzz/src/test/java/com/ezsocsi/fizzbuzz/ArgParserTest.java
@@ -1,4 +1,4 @@
-package com.ezsocsi.fizzbuzz.unittest;
+package com.ezsocsi.fizzbuzz;
 
 
 import com.ezsocsi.fizzbuzz.ArgParser;
