+++ 02-TextMunger/src/test/java/TextMungerIntegrationTest.java
@@ -1,6 +1,3 @@
-package com.ezsocsi.textmunger.integrationtest;
-
-
 import com.ezsocsi.textmunger.Main;
 
 import com.ezsocsi.utils.IntegrationTestUtils;
@@ -11,7 +8,7 @@ import java.util.regex.Pattern;
 import static org.testng.Assert.assertTrue;
 import static org.testng.Assert.assertEquals;
 
-public class MainTest extends IntegrationTestUtils {
+public class TextMungerIntegrationTest extends IntegrationTestUtils {
     @Test
     public void prints_error_when_no_arguments() {
         Main.main();
